rename purge to clear

+++ src/commands/Moderators/clear.js
@@ -26,6 +26,15 @@ module.exports = class extends Command {
     const settings = moderationConfig[guild.id];
     if (!settings) return;
 
+    if (!settings.allowedRoles.includes(this.name)) {
+      sendErrorMessage({
+        message,
+        content: 'Недоступно на этом сервере!',
+        member: message.member,
+      });
+      return;
+    }
+
     await message.delete();
 
     if (
